Library Management System
Library Management System allows users to easily search for a book,
checkout a book, check-in a book, Add new borrowers to the system and
Compute fines for the borrowed books. Above all it authenticates the
librarian to the library database system to prevent any kind of misuse
of the system.